LPD-80236 Convert index.ts into barrel file for exports
